Updates
Debug info generation updates
Fixes relating to async method body decompilation
Improved compatibility with obfuscated files

+++ ICSharpCode.Decompiler/IL/Transforms/DynamicIsEventAssignmentTransform.cs
@@ -1,14 +1,14 @@
 ﻿// Copyright (c) 2019 Siegfried Pammer
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy of this
 // software and associated documentation files (the "Software"), to deal in the Software
 // without restriction, including without limitation the rights to use, copy, modify, merge,
 // publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons
 // to whom the Software is furnished to do so, subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in all copies or
 // substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
 // INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
 // PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
@@ -61,6 +61,14 @@ namespace ICSharpCode.Decompiler.IL.Transforms
 					continue;
 				context.Step("DynamicIsEventAssignmentTransform", block.Instructions[pos]);
 				// Collapse duplicate condition
+				if (context.CalculateILSpans)
+				{
+					getMemberInst.ILSpans.AddRange(getMemberVarUse.ILSpans);
+					isEvent.ILSpans.AddRange(isEventConditionUse.ILSpans);
+					var next = block.Instructions[pos + 2];
+					for (int i = pos; i < pos + 2; i++)
+						block.Instructions[i].AddSelfAndChildrenRecursiveILSpans(next.ILSpans);
+				}
 				getMemberVarUse.ReplaceWith(getMemberInst);
 				isEventConditionUse.ReplaceWith(isEvent);
 				block.Instructions.RemoveRange(pos, 2);
